% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd2gcj.R
\name{bd2gcj}
\alias{bd2gcj}
\title{bd2gcj}
\usage{
bd2gcj(lng, lat)
}
\arguments{
\item{lng}{BD-09 longitude}

\item{lat}{BD-09 latitude}
}
\value{
location
}
\description{
BD-09 to GCJ-02
}
\details{
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 location <- bd2gcj(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
