% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcj2bd.R
\name{gcj2bd}
\alias{gcj2bd}
\title{gcj2bd}
\usage{
gcj2bd(lng, lat)
}
\arguments{
\item{lng}{GCJ-02 longitude}

\item{lat}{GCJ-02 latitude}
}
\value{
location
}
\description{
GCJ-02 to BD-09
}
\details{
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 location <- gcj2bd(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
