% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcj2wgs.R
\name{gcj2wgs}
\alias{gcj2wgs}
\title{gcj2wgs}
\usage{
gcj2wgs(lng, lat)
}
\arguments{
\item{lng}{GCJ-02 longitude}

\item{lat}{GCJ-02 latitude}
}
\value{
location
}
\description{
GCJ-02 to WGS-48
}
\details{
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 location <- gcj2wgs(lng=121.89080, lat = 31.9479)
lngs <- c(121.89080, 0, "", 119.28383)
lats <- c(31.9479, 0, "", 26.293484)
location <- gcj2wgs(lng=lngs, lat = lats)
}
\author{
t.s.helianthus
}
