% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbs2crs.R
\name{lbs2crs}
\alias{lbs2crs}
\title{lbs2crs
get epsg code by longitude, only for Beijing 1954 / 3-degree Gauss-Kruger CM coordinate System now}
\usage{
lbs2crs(x, degree = 3, sys = "beijing1954")
}
\arguments{
\item{x}{longitude, you can also directly enter the spatial attribute data, and identify the center point coordinates through bbox}
}
\value{
location_crs
}
\description{
lbs2crs
get epsg code by longitude, only for Beijing 1954 / 3-degree Gauss-Kruger CM coordinate System now
}
\examples{
lbs2crs(x = 87.68)
}
\author{
t.s.helianthus
}
