% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi2wgs.R
\name{multi2wgs}
\alias{multi2wgs}
\title{gcj2wgs}
\usage{
multi2wgs(lng, lat, crs)
}
\arguments{
\item{lng}{GCJ-02, BD or WGS longitude}

\item{lat}{GCJ-02, BD or WGS latitude}
}
\value{
location
}
\description{
GCJ-02 to WGS-48
}
\details{
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
   location <- multi2wgs(lng=121.89080, lat = 31.9479, crs="GCJ")
lng <- c(121.394948, NA, 117.499596, 93.393923, 119.2943959)
lat <- c(25.2843745, NA, 25.4958935, 33.2949457, 31.5958333)
crs <- c("GCJ", "GCJ", "FALSE", "BD", "WGS")
location <- multi2wgs(lng = lng, lat = lat, crs = crs)
}
\author{
t.s.helianthus
}
