% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_china.R
\name{not_china}
\alias{not_china}
\title{not_china}
\usage{
not_china(lng, lat)
}
\arguments{
\item{lng}{query longitude}

\item{lat}{query latitude}
}
\value{
out_of_china TRUE/FALSE
}
\description{
if location out of china
}
\details{
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 out_of_china <- not_china(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
