% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformlat.R
\name{transformlat}
\alias{transformlat}
\title{transformlat}
\usage{
transformlat(lng, lat)
}
\arguments{
\item{lng}{query longitude}

\item{lat}{query latitude}
}
\value{
ret
}
\description{
process of GCJ-02 exchange to WGS-84
}
\details{
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 ret <- transformlat(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
