% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs2bd.R
\name{wgs2bd}
\alias{wgs2bd}
\title{wgs2bd}
\usage{
wgs2bd(lng, lat)
}
\arguments{
\item{lng}{WGS-84 longitude}

\item{lat}{WGS-84 latitude}
}
\value{
location
}
\description{
WGS-84 to BD-09
}
\details{
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 location <- wgs2bd(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
