% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs2wgs.R
\name{wgs2wgs}
\alias{wgs2wgs}
\title{wgs2wgs}
\usage{
wgs2wgs(lng, lat)
}
\arguments{
\item{lng}{WGS longitude}

\item{lat}{WGS latitude}
}
\value{
location
}
\description{
WGS to WGS
}
\details{
Has no practical effect, in order to batch similar transactions(multi2wgs)
BD-09 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd by BaiduMap)
GCJ-02 formulated by the Chinese State Bureau of Mapping GIS coordinate system(Uesd byAmap/QQmap/Google map)
WGS-84 is an internationally adopted geocentric coordinate system
Referencing wandergis/coordTransform_py <\url{https://github.com/wandergis/coordTransform_py}>
}
\examples{
 location <- wgs2wgs(lng=121.89080, lat = 31.9479)
}
\author{
t.s.helianthus
}
